- fixing hideRequest issue - supporting and activating async image upload to amazon

+++ venefica-webapp/src/main/java/com/venefica/dao/ImageDaoImpl.java
@@ -46,7 +46,7 @@ public class ImageDaoImpl extends DaoBase<Image> implements ImageDao {
         
         amazonUpload.upload(files, modelType, image.getImgType());
         
-        if ( imageConfig.isDeleteImages() ) {
+        if ( !amazonUpload.isAsyncUpload() && imageConfig.isDeleteImages() ) {
             imageUtils.delete(imageId, modelType);
         }
         return imageId;
